Move provider-server to relay
