+++ core/services/ocr2/delegate.go
@@ -597,7 +597,7 @@ func (d *Delegate) newServicesGenericPlugin(
 	} else {
 		d.lggr.Info("provider is not a LOOPP provider, switching to provider server")
 
-		ps, err := NewProviderServer(provider, types.OCR2PluginType(cconf.ProviderType), d.lggr)
+		ps, err := relay.NewProviderServer(provider, types.OCR2PluginType(cconf.ProviderType), d.lggr)
 		if err != nil {
 			return nil, errors.New(fmt.Sprintf("cannot start EVM provider server: %s", err))
 		}
